Migrate the dependency resolution test to the new test framework
