+++ plugin-management-library/src/test/java/io/jenkins/tools/pluginmanager/impl/PluginManagerTest.java
@@ -722,74 +722,6 @@ public class PluginManagerTest {
     }
 
     @Test
-    public void findPluginsAndDependenciesTest() throws IOException {
-        Plugin plugin1 = new Plugin("plugin1", "1.0", null, null);
-        Plugin replaced = new Plugin("replaced", "1.0", null, null);
-        Plugin plugin2 = new Plugin("plugin2", "2.0", null, null);
-        Plugin plugin3 = new Plugin("plugin3", "3.0", null, null);
-
-        Config config = Config.builder()
-                .withJenkinsWar(Settings.DEFAULT_WAR)
-                .withPluginDir(Files.createTempDirectory("tmpplugins").toFile())
-                .build();
-
-        PluginManager pluginManager = new PluginManager(config);
-        PluginManager pluginManagerSpy = spy(pluginManager);
-
-        Map<String, Plugin> plugin1Dependencies = new HashMap<>();
-        Plugin plugin1Dependency1 = new Plugin("plugin1Dependency1", "1.0.1", null, null);
-        Plugin plugin1Dependency2 = new Plugin("plugin1Dependency2", "1.0.2", null, null);
-        Plugin replaced1 = new Plugin("replaced", "1.0.1", null, null);
-
-        //Ideally if this plugin had dependencies that are different from the plugin that replaces it, they would get
-        //cleaned up - see Jira JENKINS-58775
-        Map<String, Plugin> replacedDependencies = new HashMap<>();
-
-        plugin1Dependencies.put("plugin1", plugin1);
-        plugin1Dependencies.put("plugin1Dependency1", plugin1Dependency1);
-        plugin1Dependencies.put("plugin1Dependency2", plugin1Dependency2);
-        plugin1Dependencies.put("replaced", replaced1);
-
-        Map<String, Plugin> plugin2Dependencies = new HashMap<>();
-        Plugin plugin2Dependency1 = new Plugin("plugin2Dependency1", "2.0.1", null, null);
-        Plugin plugin2Dependency2 = new Plugin("plugin2Dependency2", "2.0.2", null, null);
-        Plugin replaced2 = new Plugin("replaced", "2.0.2", null, null);
-        Plugin replacedSecond = new Plugin("replaced2", "2.0.2", null, null);
-
-        plugin2Dependencies.put("plugin2Dependency1", plugin2Dependency1);
-        plugin2Dependencies.put("plugin2Dependency2", plugin2Dependency2);
-        plugin2Dependencies.put("replaced", replaced2);
-        plugin2Dependencies.put("replaced2", replacedSecond);
-
-        Map<String, Plugin> plugin3Dependencies = new HashMap<>();
-        Plugin plugin3Dependency1 = new Plugin("plugin3Dependency1", "3.0.1", null, null);
-        Plugin replacedSecond2 = new Plugin("replaced2", "3.2", null, null);
-
-        plugin3Dependencies.put("plugin3Dependency1", plugin3Dependency1);
-        plugin3Dependencies.put("replaced2", replacedSecond2);
-
-        doReturn(plugin1Dependencies).when(pluginManagerSpy).resolveRecursiveDependencies(plugin1);
-        doReturn(replacedDependencies).when(pluginManagerSpy).resolveRecursiveDependencies(replaced);
-        doReturn(plugin2Dependencies).when(pluginManagerSpy).resolveRecursiveDependencies(plugin2);
-        doReturn(plugin3Dependencies).when(pluginManagerSpy).resolveRecursiveDependencies(plugin3);
-
-        List<Plugin> requestedPlugins = new ArrayList<>(Arrays.asList(plugin1, plugin2, plugin3, replaced));
-
-        Map<String, Plugin> pluginsAndDependenciesMap = pluginManagerSpy.findPluginsAndDependencies(requestedPlugins);
-
-        List<Plugin> pluginsAndDependencies = new ArrayList<>(pluginsAndDependenciesMap.values());
-
-        Collections.sort(pluginsAndDependencies);
-
-        List<Plugin> expectedPlugins = new ArrayList<>(Arrays.asList(plugin1, plugin1Dependency1, plugin1Dependency2,
-                plugin2Dependency1, plugin2Dependency2, plugin3Dependency1, replaced2, replacedSecond2));
-
-        Collections.sort(expectedPlugins);
-
-        assertEquals(expectedPlugins, pluginsAndDependencies);
-    }
-
-    @Test
     public void downloadPluginAlreadyInstalledTest() {
         Map<String, Plugin> installedVersions = new HashMap<>();
         installedVersions.put("plugin1", new Plugin("plugin1", "1.0", null, null));
